backend/requirements.txt
charset-normalizer>=3.4.1
google-ai-generativelanguage>=0.6.15
google-api-core>=2.24.2
google-api-python-client>=2.164.0
google-auth>=2.38.0
google-genai>=1.5.0
imageio>=2.37.0
imageio_ffmpeg>=0.6.0
moviepy>=2.1.2
pillow>=10.4.0
proglog>=0.1.10
python-dotenv>=1.0.1
azure-cognitiveservices-speech>=1.42.0
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.3.0
pydantic-settings>=2.0.3
requests>=2.31.0
redis>=4.5.4
celery>=5.2.7